requirements.txt
# streamlit
flask
flask-cors
requests
python-dotenv
groq
requests-toolbelt
pandas
numpy<2.0.0
scipy>=1.9.0,<1.13.0
scikit-learn>=1.3.0,<1.5.0
joblib>=1.3.0
xgboost>=1.7.0,<2.1.0
androguard